--- website/src/lib/theme.ts
@@ -1,98 +1,0 @@
-// We probably don't need this now that I understand svelte more.
-// Can later move this to css classes when design system is complete
-export const theme = {
-    colors: {
-        teal10: '#FEFCFB',
-        teal20: '#FEFAF7',
-        teal30: '#FAF5F1',
-        teal40: '#F7F0ED',
-        teal50: '#F1EAE7',
-        teal60: '#EBE4E1',
-        teal70: '#E6DEDB',
-        teal80: '#E7DBD5',
-        teal90: '#E4D8D2',
-        teal100: '#DED2CD',
-        yellow10: '#FFF4E0',
-        yellow20: '#FFE8C2',
-        yellow30: '#FED48B',
-        yellow40: '#FFC869',
-        yellow50: '#EBAF45',
-        yellow60: '#CE9025',
-        yellow70: '#845300',
-        yellow80: '#684000',
-        yellow90: '#462900',
-        yellow100: '#241301',
-        orange10: '#FFEEEB',
-        orange20: '#FFDDD6',
-        orange30: '#FFBCAE',
-        orange40: '#FF9985',
-        orange50: '#FB7155',
-        orange60: '#D44F34',
-        orange70: '#B6351B',
-        orange80: '#9E230A',
-        orange90: '#701300',
-        orange100: '#380900',
-        red10: '#FEEBEE',
-        red20: '#FED6D6',
-        red30: '#FEBBBE',
-        red40: '#FD979F',
-        red50: '#FB6D80',
-        red60: '#D54A63',
-        red70: '#C02649',
-        red80: '#A7133A',
-        red90: '#720929',
-        red100: '#390613',
-        purple10: '#FAF3FB',
-        purple20: '#F6E2F9',
-        purple30: '#D9C3F0',
-        purple40: '#C0A4F5',
-        purple50: '#A580F1',
-        purple60: '#9166EA',
-        purple70: '#7147CE',
-        purple80: '#57369E',
-        purple90: '#3A2174',
-        purple100: '#201340',
-        blue10: '#EBF8FF',
-        blue20: '#D6F2FF',
-        blue30: '#9FD7EF',
-        blue40: '#75BFDE',
-        blue50: '#58A1C1',
-        blue60: '#2B7FA4',
-        blue70: '#0C6387',
-        blue80: '#004968',
-        blue90: '#003044',
-        blue100: '#001E2B',
-        green10: '#F6FFF1',
-        green20: '#D1FFC2',
-        green30: '#9DEE8D',
-        green40: '#83D472',
-        green50: '#67B756',
-        green60: '#489434',
-        green70: '#317A1B',
-        green80: '#1F6507',
-        green90: '#124200',
-        green100: '#092100',
-        gray10: '#E4DCD8',
-        gray20: '#D9D2CF',
-        gray30: '#CDC8C6',
-        gray40: '#B8B1AE',
-        gray50: '#948D8A',
-        gray60: '#77706D',
-        gray70: '#5E5754',
-        gray80: '#463F3C',
-        gray90: '#302926',
-        gray100: '#17100D',
-        brown10: '#DCC3B8',
-        brown20: '#A1887F',
-        brown30: '#705854',
-        brown40: '#68524E',
-        brown50: '#594744',
-        brown60: '#4F3F40',
-        brown70: '#473636',
-        brown80: '#3C2C2D',
-        brown90: '#332425',
-        brown100: '#280E11',
-    },
-} as const;
-
-export type ThemeColors = keyof typeof theme.colors;
